requirements.txt
# Core
pydo>=0.1.0
python-dotenv>=1.0.0
boto3>=1.34.0
jinja2>=3.1.2

# API - using versions that work together
fastapi==0.100.0           # Last version with solid Pydantic v1 support
uvicorn[standard]==0.23.2
httpx==0.25.0

# Pydantic v1 - stable for Python 3.14
pydantic==1.10.13
# NO pydantic-settings - it requires v2!

# For settings management with Pydantic v1, use python-dotenv directly
# or add this simple settings class:

# Optional
requests>=2.31.0
paramiko>=3.4.0
colorama>=0.4.6
rich>=13.7.0